Set up a copy of the maven repository from argouml-downloads.tigrs.org.
I have removed the no longer used artifacts.

Most of these artifacts were put in the argouml maven repository without
understanding how the maven world works and this is not addressing that.
This is just a quick fix to keep it working without tigris. The work to
make the argouml infrastructure more maven-ish will have to continue.